# ApiDockerTecnimotors/ApiDockerTecnimotors/obj/Debug/net8.0/EndpointInfo/ApiDockerTecnimotors.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiDockerTecnimotors",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TlAuth"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TlAuth"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TlAuth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/DetailClienteUuid": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrDetailCliente"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrDetailCliente"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrDetailCliente"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CarritoCotizacion/ListadoCarritoList": {
      "post": {
        "tags": [
          "CarritoCotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrUuidCoti"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrUuidCoti"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrUuidCoti"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CarritoCotizacion/RegisterCarritoList": {
      "post": {
        "tags": [
          "CarritoCotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrModelsCarrito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrModelsCarrito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrModelsCarrito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CarritoCotizacion/RemoveCarritoList": {
      "post": {
        "tags": [
          "CarritoCotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromCarritoListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromCarritoListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromCarritoListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CarritoCotizacion/UpdateCantidadCarrito": {
      "post": {
        "tags": [
          "CarritoCotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFromCarritoModal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFromCarritoModal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFromCarritoModal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CarritoCotizacion/RegisterCotizacionList": {
      "post": {
        "tags": [
          "CarritoCotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrLoginCarrito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrLoginCarrito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrLoginCarrito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CarritoCotizacion/RegistrarCarritologin": {
      "post": {
        "tags": [
          "CarritoCotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCarritoRegister"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCarritoRegister"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ListCarritoRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListCarritoRegister": {
        "type": "object",
        "properties": {
          "listCarritoLogin": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TlModelsCarrito"
            },
            "nullable": true
          },
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "quote": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RemoveFromCarritoListRequest": {
        "type": "object",
        "properties": {
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TlAuth": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "celular": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "repassword": {
            "type": "string",
            "nullable": true
          },
          "termaccept": {
            "type": "boolean"
          },
          "fecharegistro": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TlModelsCarrito": {
        "type": "object",
        "properties": {
          "idcarritocotizacion": {
            "type": "integer",
            "format": "int32"
          },
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "ordernshopping": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "unidad": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "marcaoriginal": {
            "type": "string",
            "nullable": true
          },
          "medida": {
            "type": "string",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "medidaestandarizado": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "familia": {
            "type": "string",
            "nullable": true
          },
          "subfamilia": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "producto": {
            "type": "string",
            "nullable": true
          },
          "vendor": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "pathimagen": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "fecharegistro": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TrDetailCliente": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TrLoginCarrito": {
        "type": "object",
        "properties": {
          "listCarrito": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TlModelsCarrito"
            },
            "nullable": true
          },
          "uuidcliente": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TrModelsCarrito": {
        "type": "object",
        "properties": {
          "idcarritocotizacion": {
            "type": "integer",
            "format": "int32"
          },
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "ordernshopping": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "unidad": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "marcaoriginal": {
            "type": "string",
            "nullable": true
          },
          "medida": {
            "type": "string",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "medidaestandarizado": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "familia": {
            "type": "string",
            "nullable": true
          },
          "subfamilia": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "producto": {
            "type": "string",
            "nullable": true
          },
          "vendor": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "pathimagen": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "fecharegistro": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TrUuidCoti": {
        "type": "object",
        "properties": {
          "uuidcliente": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateFromCarritoModal": {
        "type": "object",
        "properties": {
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}